+++ CHANGES.rst
@@ -9,6 +9,8 @@ Changelog
   getHistory knows that it queries only relevant last event and when overrided,
   the package overriding it may avoid heavy processing if relevant.
   [gbastien]
+- Get rid of unittest2.
+  [gbastien]
 
 1.9 (2016-10-12)
 ----------------